Avoid repeat queues of the same waker where possible
It's entirely possible for Interest::register to be called repeatedly
with the same waker. If this happens without MassNotify::notify_all
getting called then it will result in a memory leak.

We can't entirely do away with this but we can avoid the memory leak in
cases where Interest::register is being called repeatedly with the same
waker.