Remove thread-id check in favour of keeping a watermark reference
The thread-id check doesn't really check what we want to check, which is
that the watermark for where we last added the waker hasn't changed. By
keeping a reference to the watermark of the last queue that we've seen
we can directly check whether we need to enqueue the new waker.